static/app/components/events/interfaces/crashContent/exception/utils.tsx
import React from 'react';

import type {Frame} from 'sentry/types';
import {getFileExtension} from 'sentry/utils/fileExtension';

const fileNameBlocklist = ['@webkit-masked-url'];
export function isFrameFilenamePathlike(frame: Frame): boolean {
  let filename = frame.absPath ?? '';
  try {
    filename = new URL(filename).pathname.split('/').reverse()[0];
  } catch {
    // do nothing
  }

  return (
    // If all filenames are anonymous, we do not want to show this alert
    (frame.filename === '<anonymous>' && frame.inApp) ||
    // If all function names are on the blocklist, we do not want to show this alert
    fileNameBlocklist.includes(frame.function ?? '') ||
    // If all absolute paths do not have a file extension, we do not want to show this alert
    (!!frame.absPath && !getFileExtension(filename))
  );
}

// need to add comments
export function Linkify({exceptionText}: {exceptionText?: string}): React.ReactElement {
  if (!exceptionText) {
    return <React.Fragment>{''}</React.Fragment>;
  }

  const urlRegex =
    /https?:\/\/(?:www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9]{1,6}\b(?:[-a-zA-Z0-9@:%_\+.~#?&\/=,\[\]]*)/gi;

  const parts = exceptionText.split(urlRegex);
  const urls = exceptionText.match(urlRegex);

  const elements: React.ReactNode[] = parts.flatMap((part, index) => {
    const link =
      urls && urls[index] ? (
        <a key={`link-${index}`} href={urls[index]}>
          {urls[index]}
        </a>
      ) : null;

    return [<React.Fragment key={`text-${index}`}>{part}</React.Fragment>, link];
  });

  return <React.Fragment>{elements}</React.Fragment>;
}

// Maps the SDK name to the url token for docs
export const sourceMapSdkDocsMap: Record<string, string> = {
  'sentry.javascript.browser': 'javascript',
  'sentry.javascript.node': 'node',
  'sentry.javascript.react': 'react',
  'sentry.javascript.angular': 'angular',
  'sentry.javascript.angular-ivy': 'angular',
  'sentry.javascript.ember': 'ember',
  'sentry.javascript.gatsby': 'gatsby',
  'sentry.javascript.vue': 'vue',
  'sentry.javascript.nextjs': 'nextjs',
  'sentry.javascript.remix': 'remix',
  'sentry.javascript.svelte': 'svelte',
  'sentry.javascript.sveltekit': 'sveltekit',
  'sentry.javascript.react-native': 'react-native',
  'sentry.javascript.atro': 'astro',
};
